WIP: Make repair coordination less expensive by moving MerkleTrees off heap
Fully-offheap merkle tree deserialization

This uses a buffer that we size precisely once we know how many nodes are the tree

Buildint the trees still happens on heap - not because it wouldn't benefit,
but because it would require a bit of guessing about how large to pre-size
that buffer, and I think we'd be wrong more often than we'd be right.

Also, the cost to build a single tree isn't great, but the cost to hold
RF of them on the coordinator is at least 3x worse, and potentially 10-12x
worse.

Co-authored-by: Aleksey Yeschenko <aleksey@yeschenko.com>
Co-authored-by: Jeff Jirsa <jjirsa@apple.com>