Update port when reconnecting to pre-4.0 SSL storage
On a failed outbound connection to a node with pending data, recheck
the messaging version before reattempting the connection.

Prior to this change, if the endpoint version was incorrectly set
to 4.0 when the node was running 3.0 with an SSL storage port
the connection would continuously try to reconnect on the wrong port.

The patch also improves some of the log messages to include the
actual port being connected to as well as the canonical endpoint for
the node.

Patch by Jon Meredith & Andy Tolbert; reviewed by Aleksey Yeschenko for
CASSANDRA-15727

Co-authored-by: Jon Meredith <jmeredithco@gmail.com>
Co-authored-by: Andy Tolbert <andy_tolbert@apple.com>